chore(moo): skip cascade delete tests iff no deleteOne

+++ packages/moo/test/02.plugins-cascade-delete.spec.ts
@@ -58,12 +58,15 @@ describe('Cascade Delete Plugin', () => {
 
   it('should delete refs when a file deleted', async () => {
     const file = await File.findOne({ name: 'file1' });
-    await file.deleteOne();
 
-    const refs = await Reference.find();
-    const items = await Item.find();
+    if ('deleteOne' in file) {
+      await file.deleteOne();
 
-    expect(refs.length).equal(0);
-    expect(items.length).equal(2);
+      const refs = await Reference.find();
+      const items = await Item.find();
+
+      expect(refs.length).equal(0);
+      expect(items.length).equal(2);
+    }
   });
 });
